backend/requirements.txt
# fastapi
fastapi==0.111.0
uvicorn[standard]==0.30.1

# pydantic v2
pydantic==2.7.4
pydantic-settings==2.3.3

# langchain & langgraph
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langgraph==0.2.28

# llm providers
langchain-openai==0.1.23
langchain-groq==0.1.9

# web search
duckduckgo-search==6.2.13
tavily-python==0.3.9

# utilities
python-dotenv==1.0.1
httpx==0.27.0

# langsmith tracing (optional)
langsmith==0.1.98
